Harvest any warning we had in our codebase.
Unitary tests under meson have much more strict compile flags which in
turn leads to failing compilation when there are any warnings.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-7964

+++ ctrld/mctp-ctrl.c
@@ -397,7 +397,7 @@ static int do_mctp_cmdline(const mctp_cmdline_args_t *cmd, int sock_fd)
 		}
 	}
 
-	printf("Command Done in [%lld] ms\n", (t_end - t_start));
+	printf("Command Done in [%zu] ms\n", (size_t)(t_end - t_start));
 
 	return MCTP_CMD_SUCCESS;
 }
